+++ src/Modules/SPSUserSync.Common/Public/Clear-SPSLogFolder.ps1
@@ -1,4 +1,4 @@
-function Clear-SPSLogFolder {
+﻿function Clear-SPSLogFolder {
     <#
         .SYNOPSIS
         Deletes old log files from a folder based on a retention window.