+++ src/Modules/SPSUserSync.Common/Public/Get-SPSSyncSetting.ps1
@@ -1,4 +1,4 @@
-function Get-SPSSyncSetting {
+﻿function Get-SPSSyncSetting {
     <#
         .SYNOPSIS
         Loads and caches the sync-settings.psd1 file.